Integrate provider-aware prompt caching (#3895) which touched the same
engine.py import block + the three verdict receipts. Resolution keeps both
sides: `cache_policy` (from #3895) and `failed_members` (this branch) coexist
in every receipt; the `MemberFailure` import and `failures_out` recording are
preserved.

+++ agents/health_monitor/checks/audio.py
@@ -24,6 +24,6 @@ async def check_audio_ducker_liveness() -> list[CheckResult]:
             group="audio",
             status=Status.HEALTHY,
             message="ducking on MPC hardware — software ducker retired",
-            duration_ms=(time.monotonic() - t) * 1000,
+            duration_ms=int((time.monotonic() - t) * 1000),
         )
     ]
